requirements_test.txt
django-nose
django-webtest
exam
mock
webtest
